Updating ZoomMap to make changing years faster
When the ZoomMap component receives new props, we now check to see
if we are still mapping the same granularity of data (i.e. states
vs. counties).  If so, then there is no need to clear the map,
redraw the boundaries, and recolor.  Instead, we simply update
the colors of the map.  This is substantially faster